fine tuning/requirements.txt
# Core dependencies for fine-tuning pipeline

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Training (Tinker)
# tinker  # Install via Tinker's instructions

# Inference (MLX) - Apple Silicon only
mlx>=0.15.0
mlx-lm>=0.5.0
transformers>=4.40.0

# Model utilities
peft>=0.10.0
huggingface-hub>=0.22.0

# Validation/Comparison
anthropic>=0.25.0
scikit-learn>=1.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Optional: For notebook exploration
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
